third_party/WebKit/LayoutTests/http/tests/xmlhttprequest/access-control-preflight-sync-not-supported-expected.txt

CONSOLE ERROR: line 34: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/access-control-preflight-denied-xsrf.php. No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
PASS: Request successfully blocked.
